My Sum

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,6 +1,6 @@
 #include "function_pointers.h"
 /*
- * array_iterator - executes a function given as a parameter
+ * array_iterator - a function given as a parameter
  * on each element of an array.
  * @array: an array of integers
  * @size: the size of array
@@ -9,7 +9,7 @@
  *
  * Return: nothing
  */
-void array_iterator(int *array, int size, void (*action)(int))
+void array_iterator(int *array, size_t size, void (*action)(int))
 {
 int i;
 if (!array || !size || !action)
